fix: rename package to manipcli
