+++ .gitattributes
@@ -1,0 +1,2 @@
+go.mod filter=remod
+go.sum filter=remod
