+++ go.mod
@@ -3,7 +3,7 @@ module go.aporeto.io/manipulate
 go 1.16
 
 require (
-	go.aporeto.io/elemental v1.100.1-0.20211029010343-e264f153acc2
+	go.aporeto.io/elemental v1.100.1-0.20211007231947-44c1aabe7914
 	go.aporeto.io/wsc v1.36.1-0.20210528201141-2af38def4959
 )
 
@@ -19,7 +19,6 @@ require (
 	github.com/mitchellh/copystructure v1.2.0
 	github.com/olekukonko/tablewriter v0.0.5
 	github.com/opentracing/opentracing-go v1.2.0
-	github.com/pkg/errors v0.9.1 // indirect
 	github.com/smartystreets/assertions v1.0.0 // indirect
 	github.com/smartystreets/goconvey v1.6.4
 	github.com/spf13/cobra v1.1.1
