+++ manipcli/cli.go
@@ -1,12 +1,41 @@
-package cli
+package manipcli
 
 import (
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 	"go.aporeto.io/elemental"
+	"go.uber.org/zap"
 )
 
-func GenerateCobraCommand(modelManager elemental.ModelManager) *cobra.Command {
+// generateConfig hods the generate configuration.
+type generateConfig struct {
+	ignoredIdentities map[string]struct{}
+}
+
+// GenerateOption represents an option can for the generate command.
+type GenerateOption func(generateConfig)
+
+// GenerateOptionIgnoreIdentities sets which non-private identities should be ignored.
+func GenerateOptionIgnoreIdentities(identities ...elemental.Identity) GenerateOption {
+	return func(g generateConfig) {
+
+		var m = make(map[string]struct{}, len(identities))
+		for _, i := range identities {
+			m[i.Name] = struct{}{}
+		}
+
+		g.ignoredIdentities = m
+	}
+}
+
+// GenerateCobraCommand generates the API commands and subcommands based on the model manager.
+func GenerateCobraCommand(modelManager elemental.ModelManager, manipulatorMaker ManipulatorMaker, options ...GenerateOption) *cobra.Command {
+
+	cfg := generateConfig{}
+
+	for _, opt := range options {
+		opt(cfg)
+	}
 
 	// API Command
 	APICommand := &cobra.Command{
@@ -20,16 +49,11 @@ func GenerateCobraCommand(modelManager elemental.ModelManager) *cobra.Command {
 		},
 	}
 
-	APICommand.PersistentFlags().StringP(FlagAPI, "A", "", "API to use.")
-	APICommand.PersistentFlags().Bool(FlagAPISkipVerify, false, "Skip certificates validation. This is unsecure.")
-	APICommand.PersistentFlags().StringP(FlagNamespace, "n", "", "Namespace to use.")
 	APICommand.PersistentFlags().StringP(FlagOutput, "o", "default", "Format to used print output. Options are 'table', 'json', 'yaml', 'none', 'template' or 'default'.")
 	// TODO: Manage output template
 	// APICommand.PersistentFlags().String(FlagOutputTemplate, "", "When output is set to 'template', this defines the template to use using Go format.")
-	APICommand.PersistentFlags().StringP(FlagToken, "t", "", "Token to use.")
 	APICommand.PersistentFlags().StringSliceP(FormatTypeColumn, "c", nil, "Only show the given columns. Only valid when '--output=table'.")
 	APICommand.PersistentFlags().StringSliceP(FlagParameters, "p", nil, "Additional parameter to the request, in the form of key=value.")
-	APICommand.PersistentFlags().String(FlagEncoding, "msgpack", "Encoding to use.")
 
 	// Create command
 	createCommands := &cobra.Command{
@@ -130,36 +154,54 @@ func GenerateCobraCommand(modelManager elemental.ModelManager) *cobra.Command {
 	// Generate subcommands for each identity
 	for _, identity := range modelManager.AllIdentities() {
 
+		if _, ok := cfg.ignoredIdentities[identity.Name]; ok {
+			continue
+		}
+
 		if identity.Private {
 			continue
 		}
 
-		if cmd, err := generateCreateCommandForIdentity(identity, modelManager); err == nil {
+		if cmd, err := generateCreateCommandForIdentity(identity, modelManager, manipulatorMaker); err == nil {
 			createCommands.AddCommand(cmd)
+		} else {
+			zap.L().Debug("unable to generate create command for identity", zap.String("identity", identity.Name), zap.Error(err))
 		}
 
-		if cmd, err := generateUpdateCommandForIdentity(identity, modelManager); err == nil {
+		if cmd, err := generateUpdateCommandForIdentity(identity, modelManager, manipulatorMaker); err == nil {
 			updateCommands.AddCommand(cmd)
+		} else {
+			zap.L().Debug("unable to generate update command for identity", zap.String("identity", identity.Name), zap.Error(err))
 		}
 
-		if cmd, err := generateDeleteCommandForIdentity(identity, modelManager); err == nil {
+		if cmd, err := generateDeleteCommandForIdentity(identity, modelManager, manipulatorMaker); err == nil {
 			deleteCommands.AddCommand(cmd)
+		} else {
+			zap.L().Debug("unable to generate delete command for identity", zap.String("identity", identity.Name), zap.Error(err))
 		}
 
-		if cmd, err := generateDeleteManyCommandForIdentity(identity, modelManager); err == nil {
+		if cmd, err := generateDeleteManyCommandForIdentity(identity, modelManager, manipulatorMaker); err == nil {
 			deleteManyCommands.AddCommand(cmd)
+		} else {
+			zap.L().Debug("unable to generate delete-many command for identity", zap.String("identity", identity.Name), zap.Error(err))
 		}
 
-		if cmd, err := generateGetCommandForIdentity(identity, modelManager); err == nil {
+		if cmd, err := generateGetCommandForIdentity(identity, modelManager, manipulatorMaker); err == nil {
 			getCommands.AddCommand(cmd)
+		} else {
+			zap.L().Debug("unable to generate get command for identity", zap.String("identity", identity.Name), zap.Error(err))
 		}
 
-		if cmd, err := generateListCommandForIdentity(identity, modelManager); err == nil {
+		if cmd, err := generateListCommandForIdentity(identity, modelManager, manipulatorMaker); err == nil {
 			listCommands.AddCommand(cmd)
+		} else {
+			zap.L().Debug("unable to generate list command for identity", zap.String("identity", identity.Name), zap.Error(err))
 		}
 
-		if cmd, err := generateCountCommandForIdentity(identity, modelManager); err == nil {
+		if cmd, err := generateCountCommandForIdentity(identity, modelManager, manipulatorMaker); err == nil {
 			countCommands.AddCommand(cmd)
+		} else {
+			zap.L().Debug("unable to generate count command for identity", zap.String("identity", identity.Name), zap.Error(err))
 		}
 	}
