+++ manipcli/delete.go
@@ -1,4 +1,4 @@
-package cli
+package manipcli
 
 import (
 	"context"
@@ -11,26 +11,19 @@ import (
 	"go.aporeto.io/manipulate"
 )
 
-// generateCreateCommandForIdentity generates the command to create an object based on its identity.
-func generateCreateCommandForIdentity(identity elemental.Identity, modelManager elemental.ModelManager) (*cobra.Command, error) {
+// generateDeleteCommandForIdentity generates the command to delete an object based on its identity.
+func generateDeleteCommandForIdentity(identity elemental.Identity, modelManager elemental.ModelManager, manipulatorMaker ManipulatorMaker) (*cobra.Command, error) {
 
 	cmd := &cobra.Command{
-		Use:     identity.Name,
+		Use:     fmt.Sprintf("%s <id-or-name>", identity.Name),
 		Aliases: []string{identity.Category},
-		Short:   "Allows to create a new " + identity.Name,
+		Short:   "Allows to delete an existing " + identity.Name,
+		Args:    cobra.ExactArgs(1),
 		// Aliases: TODO: Missing alias from the spec file -> To be stored in the identity ?,
 		PersistentPreRunE: persistentPreRunE,
 		RunE: func(cmd *cobra.Command, args []string) error {
 
-			manipulator, err := prepareManipulator(
-				viper.GetString(FlagAPI),
-				viper.GetString(FlagToken),
-				viper.GetString(FlagAppCredentials),
-				viper.GetString(FlagNamespace),
-				viper.GetString(FlagCACertPath),
-				viper.GetBool(FlagAPISkipVerify),
-				viper.GetString(FlagEncoding),
-			)
+			manipulator, err := manipulatorMaker()
 			if err != nil {
 				return fmt.Errorf("unable to prepare manipulator: %w", err)
 			}
@@ -46,17 +39,18 @@ func generateCreateCommandForIdentity(identity elemental.Identity, modelManager 
 				manipulate.ContextOptionFields(viper.GetStringSlice(FormatTypeColumn)),
 			}
 
-			identifiable := modelManager.IdentifiableFromString(identity.Name)
-			if err := readViperFlags(identifiable); err != nil {
-				return fmt.Errorf("unable to read flags: %w", err)
-			}
-
 			ctx, cancel := context.WithTimeout(context.Background(), 20*time.Second)
 			defer cancel()
 
 			mctx := manipulate.NewContext(ctx, options...)
-			if err := manipulator.Create(mctx, identifiable); err != nil {
-				return fmt.Errorf("unable to create %s: %w", identity.Name, err)
+
+			identifiable, err := retrieveObjectByIDOrByName(mctx, manipulator, identity, args[0], modelManager)
+			if err != nil {
+				return fmt.Errorf("unable to retrieve %s: %w", identity.Name, err)
+			}
+
+			if err := manipulator.Delete(mctx, identifiable); err != nil {
+				return fmt.Errorf("unable to delete %s: %w", identity.Name, err)
 			}
 
 			output := viper.GetString(FlagOutput)
@@ -78,11 +72,6 @@ func generateCreateCommandForIdentity(identity elemental.Identity, modelManager 
 			fmt.Println(result)
 			return nil
 		},
-	}
-
-	identifiable := modelManager.IdentifiableFromString(identity.Name)
-	if err := setViperFlags(cmd, identifiable, true); err != nil {
-		return nil, fmt.Errorf("unable to set flags for %s: %w", identity.Name, err)
 	}
 
 	return cmd, nil
